updated with claude based on requests
can you add roxygen2 code for the following function? can you also add a check if the environment already exists and if so that it just sets it to that environment and does not create a new one? can you also check if mlflow can connect to the remote server? you need this line for that: mlflow_set_tracking_uri("http://ec2-3-255-204-156.eu-west-1.compute.amazonaws.com:5000/"). can you also check if the required underlying python packages are available in the current environment?